feat: :recycle: refactor module appointments

+++ src/modules/appointment/service/appointment.service.ts
@@ -1,9 +1,9 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { BadRequestException, HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { AppointmentStatus, PaymentStatus } from '@prisma/client';
 import * as dayjs from 'dayjs';
 import { PaymentService } from 'src/modules/payment/service/payment.service';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { AppointmentCreateProps, AppointmentDeleteProps, AppointmentGetOneProps, AppointmentUpdateProps } from '../interface/appointment.interface';
+import { AppointmentCreateProps, AppointmentDeleteProps, AppointmentDeleteSchema, AppointmentGetProps, AppointmentGetSchema, AppointmentUpdateProps, AppointmentUpdateSchema } from '../interface/appointment.interface';
 
 @Injectable()
 export class AppointmentsService {
@@ -12,19 +12,19 @@ export class AppointmentsService {
     private paymentsService: PaymentService
   ) { }
 
-  async createAppointment({ userId, date, workId, email }: AppointmentCreateProps) {
-    if (!date) {
+  async createAppointment(data: AppointmentCreateProps) {
+    if (!data.date) {
       throw new HttpException('Please provide a date', HttpStatus.BAD_REQUEST);
     }
-    if (!workId) {
+    if (!data.workId) {
       throw new HttpException('The work ID (workId) is required', HttpStatus.BAD_REQUEST);
     }
 
     const existingAppointments = await this.prisma.appointment.findMany({
       where: {
         date: {
-          gte: dayjs(date).subtract(0.9, 'hour').toDate(),
-          lte: dayjs(date).add(0.9, 'hour').toDate(),
+          gte: dayjs(data.date).subtract(0.9, 'hour').toDate(),
+          lte: dayjs(data.date).add(0.9, 'hour').toDate(),
         },
         status: { not: AppointmentStatus.CANCELADO },
       },
@@ -34,7 +34,7 @@ export class AppointmentsService {
       throw new HttpException('There is already an appointment within the one-hour interval.', HttpStatus.BAD_REQUEST);
     }
 
-    const work = await this.prisma.work.findUnique({ where: { id: workId } });
+    const work = await this.prisma.work.findUnique({ where: { id: data.workId } });
     if (!work) {
       throw new HttpException('Work (workId) not found.', HttpStatus.NOT_FOUND);
     }
@@ -42,21 +42,21 @@ export class AppointmentsService {
     const appointment = await this.prisma.appointment.create({
       data: {
         title: work.name,
-        userId,
-        date,
-        workId,
+        userId: data.userId,
+        date: data.date,
+        workId: data.workId,
         paymentStatus: PaymentStatus.PENDENTE,
       },
     });
 
     try {
       const payment = await this.paymentsService.createPayment({
-        userId,
+        userId: data.userId,
         external_reference: appointment.id,
         transactionAmount: work.price,
         description: work.name,
         paymentMethodId: 'pix',
-        email,
+        email: data.email,
       });
 
       return { appointment, payment };
@@ -66,19 +66,74 @@ export class AppointmentsService {
     }
   }
 
-  async deleteAppointment({ id }: AppointmentDeleteProps) {
-    return this.prisma.appointment.delete({ where: { id } });
+  async deleteAppointment(data: AppointmentDeleteProps) {
+    const parsedData = AppointmentDeleteSchema.safeParse(data)
+    if (!parsedData.success) {
+      throw new BadRequestException(parsedData.error.errors);
+    }
+
+    const { id } = parsedData.data;
+    const existingAppointment = await this.prisma.appointment.findUnique({ where: { id } })
+
+    if (!existingAppointment) {
+      throw new NotFoundException('Compromisso não encontrado');
+    }
+
+    await this.prisma.appointment.delete({ where: { id } });
+    return { message: 'Compromisso deletado com sucesso' };
   }
 
   async getAppointments() {
     return this.prisma.appointment.findMany();
   }
 
-  async getFilteredAppointments(filters: AppointmentGetOneProps) {
-    return this.prisma.appointment.findMany({ where: filters });
+  async getFilteredAppointments(filters: AppointmentGetProps) {
+    const parsedData = AppointmentGetSchema.safeParse(filters);
+    if (!parsedData.success) {
+      throw new BadRequestException(parsedData.error.errors);
+    }
+
+    const where: any = {};
+
+    if (parsedData.data.userId) {
+      where.userId = parsedData.data.userId;
+    }
+
+    if (parsedData.data.date) {
+      const rawDate = parsedData.data.date;
+
+      const dateValue = dayjs(rawDate).subtract(-1, 'day').startOf('day'); 
+
+      if (!dateValue.isValid()) {
+        console.log('Data inválida:', rawDate);
+        throw new BadRequestException('Data inválida');
+      }
+
+      const startOfDay = dateValue.toDate(); 
+      const endOfDay = dateValue.endOf('day').toDate();  
+
+      where.date = {
+        gte: startOfDay,
+        lte: endOfDay,
+      };
+    }
+
+    return this.prisma.appointment.findMany({ where });
   }
 
-  async updateAppointment({ id, date, status, paymentStatus, paidAmount }: AppointmentUpdateProps) {
+  async updateAppointment(data: AppointmentUpdateProps) {
+    const parsedData = AppointmentUpdateSchema.safeParse(data);
+    if (!parsedData.success) {
+      throw new BadRequestException(parsedData.error.errors);
+    }
+
+    const { id, date, status, paymentStatus, paidAmount } = parsedData.data;
+    const existingAppointment = await this.prisma.appointment.findUnique({ where: { id } })
+
+    if (!existingAppointment) {
+      throw new NotFoundException('Compromisso não encontrado');
+    }
+
     return this.prisma.appointment.update({
       where: { id },
       data: { date, status, paymentStatus, paidAmount },
